refactored langchain/index.ts
