+++ src/langchain/SolanaCancelAllOrdersTool.ts
@@ -1,0 +1,35 @@
+import { PublicKey } from "@solana/web3.js";
+import { SolanaAgentKit } from "../agent";
+import { Tool } from "langchain/tools";
+
+export class SolanaCancelAllOrdersTool extends Tool {
+  name = "solana_cancel_all_orders";
+  description = `This tool can be used to cancel all orders from a Manifest market.
+  
+    Input ( input is a JSON string ):
+    marketId: string, eg "ENhU8LsaR7vDD2G1CsWcsuSGNrih9Cv5WZEk7q9kPapQ" for SOL/USDC (required)`;
+
+  constructor(private solanaKit: SolanaAgentKit) {
+    super();
+  }
+
+  protected async _call(input: string): Promise<string> {
+    try {
+      const marketId = new PublicKey(input.trim());
+      const tx = await this.solanaKit.cancelAllOrders(marketId);
+
+      return JSON.stringify({
+        status: "success",
+        message: "Cancel orders successfully",
+        transaction: tx,
+        marketId,
+      });
+    } catch (error: any) {
+      return JSON.stringify({
+        status: "error",
+        message: error.message,
+        code: error.code || "UNKNOWN_ERROR",
+      });
+    }
+  }
+}
