+++ src/langchain/SolanaCloseEmptyTokenAccounts.ts
@@ -1,0 +1,30 @@
+import { Tool } from "langchain/tools";
+import { SolanaAgentKit } from "../agent";
+
+export class SolanaCloseEmptyTokenAccounts extends Tool {
+  name = "close_empty_token_accounts";
+  description = `Close all empty spl-token accounts and reclaim the rent`;
+
+  constructor(private solanaKit: SolanaAgentKit) {
+    super();
+  }
+
+  protected async _call(): Promise<string> {
+    try {
+      const { signature, size } =
+        await this.solanaKit.closeEmptyTokenAccounts();
+
+      return JSON.stringify({
+        status: "success",
+        message: `${size} accounts closed successfully. ${size === 48 ? "48 accounts can be closed in a single transaction try again to close more accounts" : ""}`,
+        signature,
+      });
+    } catch (error: any) {
+      return JSON.stringify({
+        status: "error",
+        message: error.message,
+        code: error.code || "UNKNOWN_ERROR",
+      });
+    }
+  }
+}
