+++ src/langchain/SolanaCompressedAirdropTool.ts
@@ -1,0 +1,46 @@
+import { Tool } from "langchain/tools";
+import { SolanaAgentKit } from "../agent";
+
+export class SolanaCompressedAirdropTool extends Tool {
+  name = "solana_compressed_airdrop";
+  description = `Airdrop SPL tokens with ZK Compression (also called as airdropping tokens)
+
+  Inputs (input is a JSON string):
+  mintAddress: string, the mint address of the token, e.g., "JUPyiwrYJFskUPiHa7hkeR8VUtAeFoSYbKedZNsDvCN" (required)
+  amount: number, the amount of tokens to airdrop per recipient, e.g., 42 (required)
+  decimals: number, the decimals of the token, e.g., 6 (required)
+  recipients: string[], the recipient addresses, e.g., ["1nc1nerator11111111111111111111111111111111"] (required)
+  priorityFeeInLamports: number, the priority fee in lamports. Default is 30_000. (optional)
+  shouldLog: boolean, whether to log progress to stdout. Default is false. (optional)`;
+
+  constructor(private solanaKit: SolanaAgentKit) {
+    super();
+  }
+
+  protected async _call(input: string): Promise<string> {
+    try {
+      const parsedInput = JSON.parse(input);
+
+      const txs = await this.solanaKit.sendCompressedAirdrop(
+        parsedInput.mintAddress,
+        parsedInput.amount,
+        parsedInput.decimals,
+        parsedInput.recipients,
+        parsedInput.priorityFeeInLamports || 30_000,
+        parsedInput.shouldLog || false,
+      );
+
+      return JSON.stringify({
+        status: "success",
+        message: `Airdropped ${parsedInput.amount} tokens to ${parsedInput.recipients.length} recipients.`,
+        transactionHashes: txs,
+      });
+    } catch (error: any) {
+      return JSON.stringify({
+        status: "error",
+        message: error.message,
+        code: error.code || "UNKNOWN_ERROR",
+      });
+    }
+  }
+}
