+++ src/langchain/SolanaDeployCollectionTool.ts
@@ -1,0 +1,37 @@
+import { Tool } from "langchain/tools";
+import { SolanaAgentKit } from "../agent";
+
+export class SolanaDeployCollectionTool extends Tool {
+  name = "solana_deploy_collection";
+  description = `Deploy a new NFT collection on Solana blockchain.
+  
+    Inputs (input is a JSON string):
+    name: string, eg "My Collection" (required)
+    uri: string, eg "https://example.com/collection.json" (required)
+    royaltyBasisPoints?: number, eg 500 for 5% (optional)`;
+
+  constructor(private solanaKit: SolanaAgentKit) {
+    super();
+  }
+
+  protected async _call(input: string): Promise<string> {
+    try {
+      const parsedInput = JSON.parse(input);
+
+      const result = await this.solanaKit.deployCollection(parsedInput);
+
+      return JSON.stringify({
+        status: "success",
+        message: "Collection deployed successfully",
+        collectionAddress: result.collectionAddress.toString(),
+        name: parsedInput.name,
+      });
+    } catch (error: any) {
+      return JSON.stringify({
+        status: "error",
+        message: error.message,
+        code: error.code || "UNKNOWN_ERROR",
+      });
+    }
+  }
+}
