+++ src/langchain/SolanaFetchPriceTool.ts
@@ -1,0 +1,31 @@
+import { Tool } from "langchain/tools";
+import { SolanaAgentKit } from "../agent";
+
+export class SolanaFetchPriceTool extends Tool {
+  name = "solana_fetch_price";
+  description = `Fetch the price of a given token in USDC.
+
+  Inputs:
+  - tokenId: string, the mint address of the token, e.g., "JUPyiwrYJFskUPiHa7hkeR8VUtAeFoSYbKedZNsDvCN"`;
+
+  constructor(private solanaKit: SolanaAgentKit) {
+    super();
+  }
+
+  async _call(input: string): Promise<string> {
+    try {
+      const price = await this.solanaKit.fetchTokenPrice(input.trim());
+      return JSON.stringify({
+        status: "success",
+        tokenId: input.trim(),
+        priceInUSDC: price,
+      });
+    } catch (error: any) {
+      return JSON.stringify({
+        status: "error",
+        message: error.message,
+        code: error.code || "UNKNOWN_ERROR",
+      });
+    }
+  }
+}
