+++ src/langchain/SolanaGetWalletAddressTool.ts
@@ -1,0 +1,15 @@
+import { Tool } from "langchain/tools";
+import { SolanaAgentKit } from "../agent";
+
+export class SolanaGetWalletAddressTool extends Tool {
+  name = "solana_get_wallet_address";
+  description = `Get the wallet address of the agent`;
+
+  constructor(private solanaKit: SolanaAgentKit) {
+    super();
+  }
+
+  async _call(_input: string): Promise<string> {
+    return this.solanaKit.wallet_address.toString();
+  }
+}
