+++ src/langchain/SolanaLendAssetTool.ts
@@ -1,0 +1,35 @@
+import { Tool } from "langchain/tools";
+import { SolanaAgentKit } from "../agent";
+
+export class SolanaLendAssetTool extends Tool {
+  name = "solana_lend_asset";
+  description = `Lend idle USDC for yield using Lulo. ( only USDC is supported )
+
+  Inputs (input is a json string):
+  amount: number, eg 1, 0.01 (required)`;
+
+  constructor(private solanaKit: SolanaAgentKit) {
+    super();
+  }
+
+  async _call(input: string): Promise<string> {
+    try {
+      const amount = JSON.parse(input).amount || input;
+
+      const tx = await this.solanaKit.lendAssets(amount);
+
+      return JSON.stringify({
+        status: "success",
+        message: "Asset lent successfully",
+        transaction: tx,
+        amount,
+      });
+    } catch (error: any) {
+      return JSON.stringify({
+        status: "error",
+        message: error.message,
+        code: error.code || "UNKNOWN_ERROR",
+      });
+    }
+  }
+}
