+++ src/langchain/SolanaRaydiumCreateClmm.ts
@@ -1,0 +1,50 @@
+import { Tool } from "langchain/tools";
+import { SolanaAgentKit } from "../agent";
+import { PublicKey } from "@solana/web3.js";
+import Decimal from "decimal.js";
+import BN from "bn.js";
+
+export class SolanaRaydiumCreateClmm extends Tool {
+  name = "raydium_create_clmm";
+  description = `Concentrated liquidity market maker, custom liquidity ranges, increased capital efficiency
+  
+    Inputs (input is a json string):
+    mint1: string (required)
+    mint2: string (required)
+    configId: string (required) stores pool info, id, index, protocolFeeRate, tradeFeeRate, tickSpacing, fundFeeRate
+    initialPrice: number, eg: 123.12 (required)
+    startTime: number(seconds), eg: now number or zero (required)
+    `;
+
+  constructor(private solanaKit: SolanaAgentKit) {
+    super();
+  }
+
+  async _call(input: string): Promise<string> {
+    try {
+      const inputFormat = JSON.parse(input);
+
+      const tx = await this.solanaKit.raydiumCreateClmm(
+        new PublicKey(inputFormat.mint1),
+        new PublicKey(inputFormat.mint2),
+
+        new PublicKey(inputFormat.configId),
+
+        new Decimal(inputFormat.initialPrice),
+        new BN(inputFormat.startTime),
+      );
+
+      return JSON.stringify({
+        status: "success",
+        message: "Raydium clmm pool created successfully",
+        transaction: tx,
+      });
+    } catch (error: any) {
+      return JSON.stringify({
+        status: "error",
+        message: error.message,
+        code: error.code || "UNKNOWN_ERROR",
+      });
+    }
+  }
+}
