+++ src/langchain/SolanaRaydiumCreateCpmm.ts
@@ -1,0 +1,52 @@
+import { Tool } from "langchain/tools";
+import { SolanaAgentKit } from "../agent";
+import { PublicKey } from "@solana/web3.js";
+import BN from "bn.js";
+
+export class SolanaRaydiumCreateCpmm extends Tool {
+  name = "raydium_create_cpmm";
+  description = `Raydium's newest CPMM, does not require marketID, supports Token 2022 standard
+  
+    Inputs (input is a json string):
+    mint1: string (required)
+    mint2: string (required)
+    configId: string (required), stores pool info, index, protocolFeeRate, tradeFeeRate, fundFeeRate, createPoolFee
+    mintAAmount: number(int), eg: 1111 (required)
+    mintBAmount: number(int), eg: 2222 (required)
+    startTime: number(seconds), eg: now number or zero (required)
+    `;
+
+  constructor(private solanaKit: SolanaAgentKit) {
+    super();
+  }
+
+  async _call(input: string): Promise<string> {
+    try {
+      const inputFormat = JSON.parse(input);
+
+      const tx = await this.solanaKit.raydiumCreateCpmm(
+        new PublicKey(inputFormat.mint1),
+        new PublicKey(inputFormat.mint2),
+
+        new PublicKey(inputFormat.configId),
+
+        new BN(inputFormat.mintAAmount),
+        new BN(inputFormat.mintBAmount),
+
+        new BN(inputFormat.startTime),
+      );
+
+      return JSON.stringify({
+        status: "success",
+        message: "Raydium cpmm pool created successfully",
+        transaction: tx,
+      });
+    } catch (error: any) {
+      return JSON.stringify({
+        status: "error",
+        message: error.message,
+        code: error.code || "UNKNOWN_ERROR",
+      });
+    }
+  }
+}
