+++ src/langchain/SolanaRegisterDomainTool.ts
@@ -1,0 +1,54 @@
+import { Tool } from "langchain/tools";
+import { SolanaAgentKit } from "../agent";
+
+export class SolanaRegisterDomainTool extends Tool {
+  name = "solana_register_domain";
+  description = `Register a .sol domain name for your wallet.
+
+  Inputs:
+  name: string, eg "pumpfun.sol" (required)
+  spaceKB: number, eg 1 (optional, default is 1)
+  `;
+
+  constructor(private solanaKit: SolanaAgentKit) {
+    super();
+  }
+
+  private validateInput(input: any): void {
+    if (!input.name || typeof input.name !== "string") {
+      throw new Error("name is required and must be a string");
+    }
+    if (
+      input.spaceKB !== undefined &&
+      (typeof input.spaceKB !== "number" || input.spaceKB <= 0)
+    ) {
+      throw new Error("spaceKB must be a positive number when provided");
+    }
+  }
+
+  protected async _call(input: string): Promise<string> {
+    try {
+      const parsedInput = JSON.parse(input);
+      this.validateInput(parsedInput);
+
+      const tx = await this.solanaKit.registerDomain(
+        parsedInput.name,
+        parsedInput.spaceKB || 1,
+      );
+
+      return JSON.stringify({
+        status: "success",
+        message: "Domain registered successfully",
+        transaction: tx,
+        domain: `${parsedInput.name}.sol`,
+        spaceKB: parsedInput.spaceKB || 1,
+      });
+    } catch (error: any) {
+      return JSON.stringify({
+        status: "error",
+        message: error.message,
+        code: error.code || "UNKNOWN_ERROR",
+      });
+    }
+  }
+}
